+++ src/main/java/org/truffleruby/language/ImmutableRubyObject.java
@@ -70,6 +70,7 @@ public abstract class ImmutableRubyObject implements TruffleObject {
     // endregion
 
     // region InteropLibrary messages
+    // Specs for these messages are in spec/truffle/interop/matrix_spec.rb
     @ExportMessage
     public boolean hasLanguage() {
         return true;