+++ src/main/ruby/truffleruby/core/hash.rb
@@ -231,7 +231,7 @@ class Hash
     end
 
     return default unless Primitive.undefined?(default)
-    raise KeyError.new("key not found: #{key.inspect}", :receiver => self, :key => key)
+    raise KeyError.new("key not found: #{key.inspect}", receiver: self, key: key)
   end
 
   def fetch_values(*keys, &block)