+++ CHANGELOG.md
@@ -7,6 +7,8 @@ New features:
 * Add `Polyglot::InnerContext` to eval code in any available language in an inner isolated context (#2169).
 * Foreign objects now have a dynamically-generated class based on their interop traits like `ForeignArray` and are better integrated with Ruby objects (#2149).
 * Foreign arrays now have all methods of Ruby `Enumerable` and many methods of `Array` (#2149).
+* Foreign hashes now have all methods of Ruby `Enumerable` and many methods of `Hash` (#2149).
+* Foreign iterables (`InteropLibrary#hasIterator`) now have all methods of Ruby `Enumerable` (#2149).
 
 Bug fixes:
 
@@ -42,6 +44,8 @@ Performance:
 
 Changes:
 
+* `foreign_object.class` on foreign objects is no longer special and uses `Kernel#class` (it used to return the `java.lang.Class` object for a Java type or `getMetaObject()`, but that is too incompatible with Ruby code).
+* `Java.import name` imports a Java class in the enclosing module instead of always as a top-level constant.
 
 # 21.2.0